requirements.txt
django
psycopg2
djangorestframework
django-crontab
black
flake8